requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.0
h11==0.14.0
idna==3.4
numpy==1.24.1
pandas==1.5.2
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
